+++ vps-install/install-go.sh
@@ -128,7 +128,7 @@ SyslogIdentifier=${GO_SERVICE_NAME}
 NoNewPrivileges=yes
 PrivateTmp=yes
 ProtectSystem=strict
-ReadWritePaths=${_APP_DIR}/logs
+ReadWritePaths=${_APP_DIR}/logs ${_APP_DIR}/uploads
 AmbientCapabilities=CAP_NET_BIND_SERVICE
 
 [Install]